+++ Unigram/Unigram/Navigation/Services/NavigationState.cs
@@ -4,7 +4,7 @@ namespace Unigram.Navigation.Services
 {
     public class NavigationState : Dictionary<string, object>
     {
-        public static NavigationState GetChatMember(long chatId, int userId)
+        public static NavigationState GetChatMember(long chatId, long userId)
         {
             return new NavigationState { { "chatId", chatId }, { "userId", userId } };
         }
